refactor(grid-pipe): consolidate grid intrinsic headers into npu/a2a3/grid_intrinsic.hpp
Merge the four include/pto/common grid headers (grid_pipe.hpp,
grid_pipe_mock_spr.hpp, grid_counter_intrinsic.hpp, grid_sram_intrinsic.hpp)
into a single include/pto/npu/a2a3/grid_intrinsic.hpp, split into four
clearly delimited sections. Repoint every include -- pto_instr.hpp, the
A2/A3 GridTPush/GridTPop/grid_pipe_runtime backends, and the
distributed_ffn_grid compute kernels and smoke tests -- at the new path.
Header-location-only change; no functional change.
